capstone_project/backend/runtime/final_agent/requirements.txt
bedrock-agentcore
strands-agents
strands-agents-tools
requests
boto3
nest-asyncio
playwright
